+++ doc/CHANGES.rst
@@ -8,6 +8,10 @@ http://docs.zope.org/zope2/releases/.
 2.13.20 (unreleased)
 --------------------
 
+- In PageTemplate.pt_errors accept the check_macro_expansion argument.
+  This is added for compatibility with zope.pagetemplate 4.0.0.  The
+  argument is ignored.  See LP #732972.
+
 - Updated distributions:
 
   - Products.ZCTextIndex = 2.13.4